Another reload of the project caused some errors. Fixed and now pushing.

+++ Asciiaria/game.cpp
@@ -111,14 +111,14 @@ int loadLevel() {
 				// If it's zero, it's air.
 			case 0:
 				p_current_tile = new Tile(AIR);
-				p_current_tile->setPosition(df::Vector(row_counter, column_counter));
+				p_current_tile->setPosition(df::Vector((float)row_counter, (float)column_counter));
 				log_manager.writeLog("This is air at (%f, %f).", p_current_tile->getPosition().getX(), p_current_tile->getPosition().getY());
 				break;
 
 				// If it's 1, it's dirt. Place the dirt!
 			case 1:
 				p_current_tile = new Tile(DIRT);
-				p_current_tile->setPosition(df::Vector(row_counter, column_counter));
+				p_current_tile->setPosition(df::Vector((float)row_counter, (float)column_counter));
 				log_manager.writeLog("This is dirt at (%f, %f).", p_current_tile->getPosition().getX(), p_current_tile->getPosition().getY());
 
 				break;
